Sample map in home_page
Removed sample map from main and added it to home_page
